requirements.txt
absl-py==0.13.0
astunparse==1.6.3
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
cmake==3.20.3
colorama==0.4.4
dlib==19.22.0
et-xmlfile==1.1.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.0.1
Flask-Cors==3.0.10
flatbuffers==1.12
gast==0.4.0
google-api-core==1.30.0
google-api-python-client==2.10.0
google-auth==1.32.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
google-spreadsheet==0.0.6
googleapis-common-protos==1.53.0
grpcio==1.34.1
gunicorn==20.1.0
h5py==3.1.0
httplib2==0.19.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
Markdown==3.3.4
MarkupSafe==2.0.1
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.1
opencv-python==4.5.2.54
opencv-python-headless==4.5.2.54
openpyxl==3.0.7
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.5
Pillow==8.2.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
termcolor==1.1.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.5
Werkzeug==2.0.1
wrapt==1.12.1
